refactor: extract shared MediaVirtuosoGrid component
